Add links for user names in admin view
Currently, these link (via jinja filter 'admin_user') to the user-mode view of
a user because we don't have an admin detail-view for a user. We should probably
create an admin user detail page. The 'Edit User Balance' and some of the editable
properties of 'Edit Users' (e.g. the name) should move to the user detail page.

These are bigger / different problems unrelated to this however, so this commit
Closes #67 and another issue can deal with the above problems.