Add 'Good Standing Discount' Feature
This is added in the backend as a `discount` field that represents
the total discount applied to a purchase transaction. The stored
amount in the transaction is the actual amount paid. This has the
least impact on the existing db infrastructure and re-calculating
subtotal when it's needed isn't too bad.

Currently, the $20 threshold and 5% discount are hard-coded. That
should probably not be the case, but it isn't the worst.